Implement backend lookup with `normalizeEmail`
