+++ backend/package.json
@@ -101,6 +101,7 @@
     "slug": "~1.1.0",
     "trunc-html": "~1.1.2",
     "uuid": "~3.3.3",
+    "validator": "^12.0.0",
     "wait-on": "~3.3.0",
     "xregexp": "^4.2.4"
   },
